+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Default sort order is now `scripts-markup-styles`. This is a breaking change.
 * Fix formatting of fenced Svelte code blocks inside markdown ([#129](https://github.com/sveltejs/prettier-plugin-svelte/issues/129))
 * Everything that is not explicitely a block element is now treated as an inline element, including components. This is a breaking change ([#159](https://github.com/sveltejs/prettier-plugin-svelte/issues/159))
+* Single quotes are no longer forced except inside quoted html attributes. This is a breaking change ([#94](https://github.com/sveltejs/prettier-plugin-svelte/issues/94))
 
 ## 1.4.2